TermInputText subclass
