+++ frontend/ui/widget/inputdialog.lua
@@ -129,6 +129,7 @@ local InputDialog = InputContainer:new{
     deny_keyboard_hiding = false, -- don't hide keyboard on tap outside
     enter_callback = nil,
     strike_callback = nil, -- call this on every keystroke
+    InputText = InputText,
     readonly = false, -- don't allow editing, will not show keyboard
     allow_newline = false, -- allow entering new lines (this disables any enter_callback)
     cursor_at_end = true, -- starts with cursor at end of text, ready for appending
@@ -280,7 +281,7 @@ function InputDialog:init()
     if not self.text_height or self.fullscreen then
         -- We need to find the best height to avoid screen overflow
         -- Create a dummy input widget to get some metrics
-        local input_widget = InputText:new{
+        local input_widget = self.InputText:new{
             text = self.fullscreen and "-" or self.input,
             input_type = self.input_type,
             face = self.input_face,
@@ -331,7 +332,7 @@ function InputDialog:init()
         -- (will work in case of re-init as these are saved by onClose()
         self._top_line_num, self._charpos = self.view_pos_callback()
     end
-    self._input_widget = InputText:new{
+    self._input_widget = self.InputText:new{
         text = self.input,
         hint = self.input_hint,
         face = self.input_face,
@@ -347,7 +348,6 @@ function InputDialog:init()
         margin = self.input_margin,
         input_type = self.input_type,
         text_type = self.text_type,
-        terminal_mode = self.terminal_mode,
         enter_callback = self.enter_callback or function()
             for _,btn_row in ipairs(self.buttons) do
                 for _,btn in ipairs(btn_row) do
